owd-flask-app/requirements.txt
bokeh==2.3.0
flask==1.1.2
gunicorn==20.0.4
pyyaml==5.4.1
